# src/mescobrad_edge/models/operation.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from mescobrad_edge.models.base_model_ import Model
from mescobrad_edge import util


class Operation(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, id: str=None, name: str=None, description: str=None, plugin_id: str=None):  # noqa: E501
        """Operation - a model defined in Swagger

        :param id: The id of this Operation.  # noqa: E501
        :type id: str
        :param name: The name of this Operation.  # noqa: E501
        :type name: str
        :param description: The description of this Operation.  # noqa: E501
        :type description: str
        :param plugin_id: The plugin_id of this Operation.  # noqa: E501
        :type plugin_id: str
        """
        self.swagger_types = {
            'id': str,
            'name': str,
            'description': str,
            'plugin_id': str
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'description': 'description',
            'plugin_id': 'plugin_id'
        }

        self._id = id
        self._name = name
        self._description = description
        self._plugin_id = plugin_id

    @classmethod
    def from_dict(cls, dikt) -> 'Operation':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Operation of this Operation.  # noqa: E501
        :rtype: Operation
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this Operation.


        :return: The id of this Operation.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this Operation.


        :param id: The id of this Operation.
        :type id: str
        """

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this Operation.


        :return: The name of this Operation.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Operation.


        :param name: The name of this Operation.
        :type name: str
        """

        self._name = name

    @property
    def description(self) -> str:
        """Gets the description of this Operation.


        :return: The description of this Operation.
        :rtype: str
        """
        return self._description

    @description.setter
    def description(self, description: str):
        """Sets the description of this Operation.


        :param description: The description of this Operation.
        :type description: str
        """

        self._description = description

    @property
    def plugin_id(self) -> str:
        """Gets the plugin_id of this Operation.


        :return: The plugin_id of this Operation.
        :rtype: str
        """
        return self._plugin_id

    @plugin_id.setter
    def plugin_id(self, plugin_id: str):
        """Sets the plugin_id of this Operation.


        :param plugin_id: The plugin_id of this Operation.
        :type plugin_id: str
        """

        self._plugin_id = plugin_id