# src/mescobrad_edge/models/plugin_configuration.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from mescobrad_edge.models.base_model_ import Model
from mescobrad_edge import util


class PluginConfiguration(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, plugin_id: str=None, configuration: object=None):  # noqa: E501
        """PluginConfiguration - a model defined in Swagger

        :param plugin_id: The plugin_id of this PluginConfiguration.  # noqa: E501
        :type plugin_id: str
        :param configuration: The configuration of this PluginConfiguration.  # noqa: E501
        :type configuration: object
        """
        self.swagger_types = {
            'plugin_id': str,
            'configuration': object
        }

        self.attribute_map = {
            'plugin_id': 'plugin_id',
            'configuration': 'configuration'
        }

        self._plugin_id = plugin_id
        self._configuration = configuration

    @classmethod
    def from_dict(cls, dikt) -> 'PluginConfiguration':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The PluginConfiguration of this PluginConfiguration.  # noqa: E501
        :rtype: PluginConfiguration
        """
        return util.deserialize_model(dikt, cls)

    @property
    def plugin_id(self) -> str:
        """Gets the plugin_id of this PluginConfiguration.


        :return: The plugin_id of this PluginConfiguration.
        :rtype: str
        """
        return self._plugin_id

    @plugin_id.setter
    def plugin_id(self, plugin_id: str):
        """Sets the plugin_id of this PluginConfiguration.


        :param plugin_id: The plugin_id of this PluginConfiguration.
        :type plugin_id: str
        """

        self._plugin_id = plugin_id

    @property
    def configuration(self) -> object:
        """Gets the configuration of this PluginConfiguration.


        :return: The configuration of this PluginConfiguration.
        :rtype: object
        """
        return self._configuration

    @configuration.setter
    def configuration(self, configuration: object):
        """Sets the configuration of this PluginConfiguration.


        :param configuration: The configuration of this PluginConfiguration.
        :type configuration: object
        """

        self._configuration = configuration